backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
decorator==5.1.1
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
faiss-cpu==1.7.4
Faker==19.2.0
favicon==0.7.0
fonttools==4.42.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
htbuilder==0.6.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
langchain==0.0.154
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==10.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.4
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pkgutil-resolve-name==1.3.10
protobuf==3.20.3
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.1b0
Pygments==2.16.1
pymdown-extensions==10.1
Pympler==1.0.1
pyparsing==3.0.9
pypdf2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
semver==3.0.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.4
st-annotated-text==4.0.0
streamlit==1.18.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.0
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
validators==0.20.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.16.2
